Split metadata filter utilities
