+++ packages/twenty-server/src/engine/metadata-modules/index-metadata/dtos/index-filter.input.ts
@@ -2,7 +2,7 @@ import { Field, InputType } from '@nestjs/graphql';
 
 import { BooleanFieldComparisonInput } from 'src/engine/metadata-modules/pagination/dtos/boolean-field-comparison.input';
 import { UUIDFilterComparisonInput } from 'src/engine/metadata-modules/pagination/dtos/uuid-filter-comparison.input';
-import { type MetadataFilterColumn } from 'src/engine/metadata-modules/pagination/utils/apply-metadata-filter-to-query-builder.util';
+import { type MetadataFilterColumn } from 'src/engine/metadata-modules/pagination/types/metadata-filter-column.type';
 
 @InputType('IndexFilter')
 export class IndexFilterInput {
