+++ packages/twenty-server/src/engine/metadata-modules/pagination/types/metadata-filter-shape.type.ts
@@ -1,0 +1,4 @@
+export type MetadataFilterShape<TFilter> = {
+  and?: TFilter[];
+  or?: TFilter[];
+};
