+++ packages/twenty-server/src/engine/metadata-modules/pagination/utils/__tests__/apply-metadata-filter-to-query-builder.util.spec.ts
@@ -1,10 +1,7 @@
 import { Brackets, type WhereExpressionBuilder } from 'typeorm';
 
 import { UserInputError } from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
-import {
-  applyMetadataFilterToItems,
-  applyMetadataFilterToQueryBuilder,
-} from 'src/engine/metadata-modules/pagination/utils/apply-metadata-filter-to-query-builder.util';
+import { applyMetadataFilterToQueryBuilder } from 'src/engine/metadata-modules/pagination/utils/apply-metadata-filter-to-query-builder.util';
 
 class FakeWhereBuilder {
   conditions: string[] = [];
@@ -174,18 +171,9 @@ describe('applyMetadataFilterToQueryBuilder', () => {
     expect(() => applyFilter({ id: { is: true } })).toThrow(UserInputError);
   });
 
-  it('normalizes UUID operands for database and in-memory filters', () => {
-    type UuidFilter = {
-      and?: UuidFilter[];
-      or?: UuidFilter[];
-      id?: { eq?: string; in?: string[] };
-    };
-
+  it('normalizes UUID operands for database filters', () => {
     const lowerCaseUuid = 'aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa';
     const upperCaseUuid = lowerCaseUuid.toUpperCase();
-    const filter: UuidFilter = {
-      id: { eq: upperCaseUuid, in: [upperCaseUuid] },
-    };
     const whereBuilder = applyFilter({
       id: { eq: upperCaseUuid, in: [upperCaseUuid] },
     });
@@ -194,69 +182,6 @@ describe('applyMetadataFilterToQueryBuilder', () => {
       metadataFilterParameter0: lowerCaseUuid,
       metadataFilterParameter1: [lowerCaseUuid],
     });
-    expect(
-      applyMetadataFilterToItems({
-        items: [{ id: lowerCaseUuid }],
-        filter,
-        columnByFilterField: {
-          id: { column: 'id', type: 'uuid' },
-        },
-      }),
-    ).toEqual([{ id: lowerCaseUuid }]);
-  });
-
-  it('applies the same filter semantics to batched in-memory relations', () => {
-    type TestFilter = {
-      and?: TestFilter[];
-      id?: { gt?: string; lt?: string };
-      isActive?: { is?: boolean };
-    };
-
-    const items = [
-      { id: 'a', isActive: true },
-      { id: 'b', isActive: false },
-      { id: 'c', isActive: true },
-    ];
-
-    expect(
-      applyMetadataFilterToItems<(typeof items)[number], TestFilter>({
-        items,
-        filter: {
-          and: [{ isActive: { is: true } }],
-          id: { gt: 'a', lt: 'c' },
-        },
-        columnByFilterField: {
-          id: { column: 'id', type: 'uuid' },
-          isActive: { column: 'isActive', type: 'boolean' },
-        },
-      }),
-    ).toEqual([]);
-  });
-
-  it('matches uuid filters regardless of operand casing', () => {
-    type UuidFilter = {
-      and?: UuidFilter[];
-      or?: UuidFilter[];
-      id?: { eq?: string; neq?: string; in?: string[]; notIn?: string[] };
-    };
-
-    const items = [{ id: '00000000-0000-4000-8000-00000000000a' }];
-    const columnByFilterField = {
-      id: { column: 'id', type: 'uuid' },
-    } as const;
-    const upperCasedId = items[0].id.toUpperCase();
-
-    const matching = (filter: UuidFilter) =>
-      applyMetadataFilterToItems<(typeof items)[number], UuidFilter>({
-        items,
-        filter,
-        columnByFilterField,
-      });
-
-    expect(matching({ id: { eq: upperCasedId } })).toEqual(items);
-    expect(matching({ id: { in: [upperCasedId] } })).toEqual(items);
-    expect(matching({ id: { neq: upperCasedId } })).toEqual([]);
-    expect(matching({ id: { notIn: [upperCasedId] } })).toEqual([]);
   });
 
   it('ignores undefined filter values', () => {
