+++ packages/twenty-server/src/engine/metadata-modules/pagination/utils/apply-metadata-filter-to-query-builder.util.ts
@@ -2,35 +2,10 @@ import { isDefined } from 'twenty-shared/utils';
 import { Brackets, type WhereExpressionBuilder } from 'typeorm';
 
 import { UserInputError } from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
-
-export type MetadataFilterComparison = {
-  eq?: unknown;
-  neq?: unknown;
-  gt?: unknown;
-  gte?: unknown;
-  lt?: unknown;
-  lte?: unknown;
-  like?: unknown;
-  notLike?: unknown;
-  iLike?: unknown;
-  notILike?: unknown;
-  in?: unknown[];
-  notIn?: unknown[];
-  is?: boolean | null;
-  isNot?: boolean | null;
-};
-
-// invertBooleanValues supports deprecated boolean aliases whose backing
-// column has the opposite meaning (isUIReadOnly filters run against
-// isUIEditable, since the legacy column is no longer written).
-export type MetadataFilterColumn =
-  | { column: string; type: 'uuid' }
-  | { column: string; type: 'boolean'; invertBooleanValues?: true };
-
-type MetadataFilterShape<TFilter> = {
-  and?: TFilter[];
-  or?: TFilter[];
-};
+import { type MetadataFilterColumn } from 'src/engine/metadata-modules/pagination/types/metadata-filter-column.type';
+import { type MetadataFilterComparison } from 'src/engine/metadata-modules/pagination/types/metadata-filter-comparison.type';
+import { type MetadataFilterShape } from 'src/engine/metadata-modules/pagination/types/metadata-filter-shape.type';
+import { normalizeMetadataFilterComparison } from 'src/engine/metadata-modules/pagination/utils/normalize-metadata-filter-comparison.util';
 
 type ParameterCounter = { value: number };
 
@@ -44,89 +19,6 @@ const buildBooleanCondition = (
   }
 
   return `${column} ${operator} ${value ? 'TRUE' : 'FALSE'}`;
-};
-
-const invertBooleanComparisonValue = (
-  value: boolean | null | undefined,
-): boolean | null | undefined => {
-  if (typeof value === 'boolean') {
-    return !value;
-  }
-
-  return value;
-};
-
-const normalizeMetadataComparison = ({
-  comparison,
-  column,
-}: {
-  comparison: MetadataFilterComparison;
-  column: MetadataFilterColumn;
-}): MetadataFilterComparison => {
-  const normalizedComparison = { ...comparison };
-
-  if (column.type === 'boolean' && column.invertBooleanValues) {
-    normalizedComparison.is = invertBooleanComparisonValue(comparison.is);
-    normalizedComparison.isNot = invertBooleanComparisonValue(comparison.isNot);
-  }
-
-  if (column.type === 'uuid') {
-    if (
-      (normalizedComparison.is !== undefined &&
-        normalizedComparison.is !== null) ||
-      (normalizedComparison.isNot !== undefined &&
-        normalizedComparison.isNot !== null)
-    ) {
-      throw new UserInputError(
-        'UUID is/isNot comparisons only support null values',
-      );
-    }
-
-    // Postgres compares uuid columns canonically, but the in-memory matcher
-    // compares raw strings, so operands are lowercased to keep both paths
-    // agreeing on differently-cased but equal ids.
-    return lowercaseComparisonOperands(normalizedComparison);
-  }
-
-  return normalizedComparison;
-};
-
-const lowercaseValue = (value: unknown): unknown =>
-  typeof value === 'string' ? value.toLowerCase() : value;
-
-const lowercaseComparisonOperands = (
-  comparison: MetadataFilterComparison,
-): MetadataFilterComparison => {
-  const lowercasedComparison: MetadataFilterComparison = { ...comparison };
-
-  for (const operand of [
-    'eq',
-    'neq',
-    'gt',
-    'gte',
-    'lt',
-    'lte',
-    'like',
-    'notLike',
-    'iLike',
-    'notILike',
-  ] as const) {
-    if (isDefined(lowercasedComparison[operand])) {
-      lowercasedComparison[operand] = lowercaseValue(
-        lowercasedComparison[operand],
-      );
-    }
-  }
-
-  for (const operand of ['in', 'notIn'] as const) {
-    const values = lowercasedComparison[operand];
-
-    if (isDefined(values)) {
-      lowercasedComparison[operand] = values.map(lowercaseValue);
-    }
-  }
-
-  return lowercasedComparison;
 };
 
 const applyComparisonToQueryBuilder = ({
@@ -291,7 +183,7 @@ export const applyMetadataFilterToQueryBuilder = <
     }
 
     const { column } = filterColumn;
-    const comparison = normalizeMetadataComparison({
+    const comparison = normalizeMetadataFilterComparison({
       comparison: filterValue as MetadataFilterComparison,
       column: filterColumn,
     });
@@ -304,216 +196,3 @@ export const applyMetadataFilterToQueryBuilder = <
     });
   }
 };
-
-const matchesSqlLikePattern = (value: string, pattern: string): boolean => {
-  const escapedPattern = pattern.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  const regularExpressionPattern = escapedPattern
-    .replace(/%/g, '.*')
-    .replace(/_/g, '.');
-
-  return new RegExp(`^${regularExpressionPattern}$`).test(value);
-};
-
-const matchesMetadataComparison = ({
-  value,
-  comparison,
-  column,
-}: {
-  value: unknown;
-  comparison: MetadataFilterComparison;
-  column: MetadataFilterColumn;
-}): boolean => {
-  const normalizedComparison = normalizeMetadataComparison({
-    comparison,
-    column,
-  });
-  const normalizedValue =
-    column.type === 'uuid' ? lowercaseValue(value) : value;
-
-  const stringValue =
-    typeof normalizedValue === 'string' ? normalizedValue : undefined;
-  const simpleComparisons: [
-    keyof MetadataFilterComparison,
-    (comparisonValue: unknown) => boolean,
-  ][] = [
-    ['eq', (comparisonValue) => normalizedValue === comparisonValue],
-    ['neq', (comparisonValue) => normalizedValue !== comparisonValue],
-    [
-      'gt',
-      (comparisonValue) =>
-        isDefined(stringValue) &&
-        typeof comparisonValue === 'string' &&
-        stringValue > comparisonValue,
-    ],
-    [
-      'gte',
-      (comparisonValue) =>
-        isDefined(stringValue) &&
-        typeof comparisonValue === 'string' &&
-        stringValue >= comparisonValue,
-    ],
-    [
-      'lt',
-      (comparisonValue) =>
-        isDefined(stringValue) &&
-        typeof comparisonValue === 'string' &&
-        stringValue < comparisonValue,
-    ],
-    [
-      'lte',
-      (comparisonValue) =>
-        isDefined(stringValue) &&
-        typeof comparisonValue === 'string' &&
-        stringValue <= comparisonValue,
-    ],
-  ];
-
-  for (const [comparisonKey, matches] of simpleComparisons) {
-    const comparisonValue = normalizedComparison[comparisonKey];
-
-    if (isDefined(comparisonValue) && !matches(comparisonValue)) {
-      return false;
-    }
-  }
-
-  const likeComparisons: [keyof MetadataFilterComparison, boolean, boolean][] =
-    [
-      ['like', false, false],
-      ['notLike', false, true],
-      ['iLike', true, false],
-      ['notILike', true, true],
-    ];
-
-  for (const [comparisonKey, caseInsensitive, negate] of likeComparisons) {
-    const comparisonValue = normalizedComparison[comparisonKey];
-
-    if (!isDefined(comparisonValue)) {
-      continue;
-    }
-
-    if (!isDefined(stringValue) || typeof comparisonValue !== 'string') {
-      return false;
-    }
-
-    const candidate = caseInsensitive ? stringValue.toLowerCase() : stringValue;
-    const pattern = caseInsensitive
-      ? comparisonValue.toLowerCase()
-      : comparisonValue;
-    const matches = matchesSqlLikePattern(candidate, pattern);
-
-    if (negate ? matches : !matches) {
-      return false;
-    }
-  }
-
-  if (
-    isDefined(normalizedComparison.in) &&
-    !normalizedComparison.in.includes(normalizedValue)
-  ) {
-    return false;
-  }
-
-  if (
-    isDefined(normalizedComparison.notIn) &&
-    normalizedComparison.notIn.includes(normalizedValue)
-  ) {
-    return false;
-  }
-
-  if (
-    normalizedComparison.is !== undefined &&
-    (normalizedComparison.is === null
-      ? normalizedValue !== null && normalizedValue !== undefined
-      : normalizedValue !== normalizedComparison.is)
-  ) {
-    return false;
-  }
-
-  if (
-    normalizedComparison.isNot !== undefined &&
-    (normalizedComparison.isNot === null
-      ? normalizedValue === null || normalizedValue === undefined
-      : normalizedValue === normalizedComparison.isNot)
-  ) {
-    return false;
-  }
-
-  return true;
-};
-
-export const applyMetadataFilterToItems = <
-  TEntity extends { id: string },
-  TFilter extends MetadataFilterShape<TFilter>,
->({
-  items,
-  filter,
-  columnByFilterField,
-}: {
-  items: TEntity[];
-  filter: TFilter;
-  columnByFilterField: Record<string, MetadataFilterColumn>;
-}): TEntity[] =>
-  items.filter((item) => {
-    for (const [filterField, filterValue] of Object.entries(filter)) {
-      if (!isDefined(filterValue)) {
-        continue;
-      }
-
-      if (filterField === 'and') {
-        if (
-          !(filterValue as TFilter[]).every(
-            (subFilter) =>
-              applyMetadataFilterToItems({
-                items: [item],
-                filter: subFilter,
-                columnByFilterField,
-              }).length === 1,
-          )
-        ) {
-          return false;
-        }
-
-        continue;
-      }
-
-      if (filterField === 'or') {
-        const subFilters = filterValue as TFilter[];
-
-        if (
-          subFilters.length > 0 &&
-          !subFilters.some(
-            (subFilter) =>
-              applyMetadataFilterToItems({
-                items: [item],
-                filter: subFilter,
-                columnByFilterField,
-              }).length === 1,
-          )
-        ) {
-          return false;
-        }
-
-        continue;
-      }
-
-      const column = columnByFilterField[filterField];
-
-      if (!isDefined(column)) {
-        throw new UserInputError(`Unknown filter field: ${filterField}`);
-      }
-
-      const value: unknown = Reflect.get(item, column.column);
-
-      if (
-        !matchesMetadataComparison({
-          value,
-          comparison: filterValue as MetadataFilterComparison,
-          column,
-        })
-      ) {
-        return false;
-      }
-    }
-
-    return true;
-  });
